separated style types for easier integration
