+++ components/layout/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 
 [dependencies]
 style = { version = "*", path = "../style" }
+style_types = { path = "../style_types" }
 dom = { version = "*", path = "../dom" }
 tree = { version = "*", path = "../tree" }
 shared = { version = "*", path = "../shared" }
