+++ components/layout/src/utils.rs
@@ -2,7 +2,7 @@ use std::rc::Rc;
 
 use css::cssom::css_rule::CSSRule;
 use dom::node::Node;
-use style::value_processing::{CSSLocation, CascadeOrigin, ContextualRule};
+use style_types::{CSSLocation, CascadeOrigin, ContextualRule};
 use test_utils::css::parse_stylesheet;
 
 use crate::layout_box::LayoutBox;
